Imported ALGLIB + Improved AutoDiff
- Initial import of the AlgLib library.
- Added support for piecewise functions in AutoDiff
- Added initial implementation of points-skeleton based cylinders (instead of fibermesh cylinders)
- Refactored gradient optimizers to reduce code duplication.
- Added TermVector3D.cs class to easily define AutoDiff based 3D vector math.

+++ AutoDiff/AutoDiff/IntPower.cs
@@ -7,14 +7,14 @@ namespace AutoDiff
 {
     public class IntPower : Term
     {
-        public IntPower(Term baseTerm, int exponent)
+        public IntPower(Term baseTerm, double exponent)
         {
             Base = baseTerm;
             Exponent = exponent;
         }
 
         public Term Base { get; private set; }
-        public int Exponent { get; private set; }
+        public double Exponent { get; private set; }
 
         public override void Accept(ITermVisitor visitor)
         {
